requirements.txt
# PAWS - Perimeter AI Wildlife Surveillance
# Python dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0

# AWS Services
boto3>=1.34.0
httpx>=0.25.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
yt-dlp>=2024.1.1
python-multipart>=0.0.6
ultralytics>=8.1.0
opencv-python-headless>=4.8.0

# Optional — Nova Act browser automation SDK
# Uncomment when SDK is publicly available on PyPI:
# nova-act>=0.1.0

# Optional — Nova Sonic bidirectional streaming SDK (requires Python 3.12+)
# Uncomment if running Python 3.12+:
# aws_sdk_bedrock_runtime>=0.4.0
